Pull request cleanup and refactoring (See [this task list comment](https://github.com/avalonmediasystem/avalon/pull/2348#issuecomment-333665809))
