+++ app/controllers/master_files_controller.rb
@@ -47,7 +47,7 @@ class MasterFilesController < ApplicationController
 
   def embed
     @master_file = MasterFile.find(params[:id])
-    if can? :read, @master_file.media_object
+    if can? :read, @master_file
       @stream_info = secure_streams(@master_file.stream_details)
     end
     respond_to do |format|
