+++ app/models/security_handler.rb
@@ -1,11 +1,13 @@
 class SecurityHandler
   class << self
     def secure_url(url, context={})
-      @shim&.call(url, context) || url
+      return @shim.call(url, context) unless @shim.nil?
+      SecurityService.new.rewrite_url(url, context)
     end
 
     def secure_cookies(context={})
-      @cookie_shim&.call(context) || {}
+      return @cookie_shim.call(context) unless @cookie_shim.nil?
+      SecurityService.new.create_cookies(context)
     end
 
     def rewrite_url(&block)
